move and rename lesson about exceptions
