+++ Lessons/Exceptions/UserSubcribtions.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace ExceptionsExtensionsEvents
+namespace Exceptions
 {
     public class UserSubcriptions
     {
