+++ SolutionForBeginners.sln
@@ -35,15 +35,15 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ToDoClient", "ToDoClient\To
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Delegates", "Delegates\Delegates.csproj", "{AA4AAFA7-4290-4DB3-A9C4-6C4DBDC7962B}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExceptionsExtensionsEvents", "ExceptionsExtensionsEvents\ExceptionsExtensionsEvents.csproj", "{C4877D7E-A746-4889-9019-99E3FB3D162D}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "CafeForDevs", "CafeForDevs", "{718A43CA-5F1F-40D3-8ECD-046470127B8A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CafeForDevs.Client", "CafeForDevs\CafeForDevs.Client\CafeForDevs.Client.csproj", "{98ABD184-AF2E-429A-9DC6-8D39827A3193}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CafeForDevs.Client", "CafeForDevs\CafeForDevs.Client\CafeForDevs.Client.csproj", "{98ABD184-AF2E-429A-9DC6-8D39827A3193}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CafeForDevs.Server", "CafeForDevs\CafeForDevs.Server\CafeForDevs.Server.csproj", "{1AFEAC36-1474-4DD6-AE13-03C30712E746}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CafeForDevs.Server", "CafeForDevs\CafeForDevs.Server\CafeForDevs.Server.csproj", "{1AFEAC36-1474-4DD6-AE13-03C30712E746}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CafeForDevs.Models", "CafeForDevs\CafeForDevs.Models\CafeForDevs.Models.csproj", "{C83DD383-411B-47AE-9D6E-F7A234ED3C6A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CafeForDevs.Models", "CafeForDevs\CafeForDevs.Models\CafeForDevs.Models.csproj", "{C83DD383-411B-47AE-9D6E-F7A234ED3C6A}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Exceptions", "Lessons\Exceptions\Exceptions.csproj", "{A76BF2D0-FC8F-4449-B895-FDCF7355935E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -111,10 +111,6 @@ Global
 		{AA4AAFA7-4290-4DB3-A9C4-6C4DBDC7962B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AA4AAFA7-4290-4DB3-A9C4-6C4DBDC7962B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AA4AAFA7-4290-4DB3-A9C4-6C4DBDC7962B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C4877D7E-A746-4889-9019-99E3FB3D162D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C4877D7E-A746-4889-9019-99E3FB3D162D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C4877D7E-A746-4889-9019-99E3FB3D162D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C4877D7E-A746-4889-9019-99E3FB3D162D}.Release|Any CPU.Build.0 = Release|Any CPU
 		{98ABD184-AF2E-429A-9DC6-8D39827A3193}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{98ABD184-AF2E-429A-9DC6-8D39827A3193}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{98ABD184-AF2E-429A-9DC6-8D39827A3193}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -127,6 +123,10 @@ Global
 		{C83DD383-411B-47AE-9D6E-F7A234ED3C6A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C83DD383-411B-47AE-9D6E-F7A234ED3C6A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C83DD383-411B-47AE-9D6E-F7A234ED3C6A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A76BF2D0-FC8F-4449-B895-FDCF7355935E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A76BF2D0-FC8F-4449-B895-FDCF7355935E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A76BF2D0-FC8F-4449-B895-FDCF7355935E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A76BF2D0-FC8F-4449-B895-FDCF7355935E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -141,11 +141,11 @@ Global
 		{36538CBD-01C3-49C9-9ECE-60E4753ADBB7} = {5ED0CD91-CF7B-4178-A48C-2D4EE97B4C56}
 		{9E0142C7-3165-42F9-9654-EDDFDC4FAA09} = {5ED0CD91-CF7B-4178-A48C-2D4EE97B4C56}
 		{AA4AAFA7-4290-4DB3-A9C4-6C4DBDC7962B} = {5ED0CD91-CF7B-4178-A48C-2D4EE97B4C56}
-		{C4877D7E-A746-4889-9019-99E3FB3D162D} = {5ED0CD91-CF7B-4178-A48C-2D4EE97B4C56}
 		{718A43CA-5F1F-40D3-8ECD-046470127B8A} = {5ED0CD91-CF7B-4178-A48C-2D4EE97B4C56}
 		{98ABD184-AF2E-429A-9DC6-8D39827A3193} = {718A43CA-5F1F-40D3-8ECD-046470127B8A}
 		{1AFEAC36-1474-4DD6-AE13-03C30712E746} = {718A43CA-5F1F-40D3-8ECD-046470127B8A}
 		{C83DD383-411B-47AE-9D6E-F7A234ED3C6A} = {718A43CA-5F1F-40D3-8ECD-046470127B8A}
+		{A76BF2D0-FC8F-4449-B895-FDCF7355935E} = {5ED0CD91-CF7B-4178-A48C-2D4EE97B4C56}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {A1E4D244-6987-464D-B47E-3F159C9C2209}
